Add singleton template class
Because I want to fight...

No but seriously, I think that singletons are okay as long as you don't build your entire engine around them.
For example: if you want to be able to globallt access some kind of clock. Then it's sort of unessesary
to pass that down into hundreds of scopes because it *might* be used somewhere.

But I'll also implement service providers later, which basically are singletons, except no one complains about them.