Rework memory tracking
Instead of overriding the 'new' operator, we instead create a macro called 'newp' that tracks allocations.
This helps in two (a million) ways. First: no more undef:ing of new operator every time we include third party libraries.
Secondly: if we know that a specific part of the code has no memory leaks, we can simply replace the 'newp' operator
with a regular 'new' so we don't have to track memory there, increasing performance in debug.